requirements-pi.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
duckdb==1.1.3
firecrawl==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
markdown-it-py==3.0.0
mdurl==0.1.2
nest_asyncio==1.6.0
numpy==2.1.3
openai==1.55.0
pandas==2.2.3
pillow==11.0.0
psycopg2-binary==2.9.10
PyAudio==0.2.14
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.9.4
six==1.16.0
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websockets==14.1